face_attendance/requirements.txt
Flask
face_recognition
opencv-python
pillow
numpy
mysql-connector-python
dlib
requests
setuptools
easydict
numpy
tqdm
torchvision
torch
opencv_python
Pillow
tensorboardX

